Add recipe ratings and star updates in RecipeDrawer
Implement favorite system with toggle API integration

Refactor recipe page structure by moving filters and RecipeDrawer into reusable components

Introduce AuthService to retrieve userId from session

Fix category and season popup visibility issue on RecipeDetailPage

Add recipe stats and favorite indicator on RecipeDetailPage

Fix ingredient unit removal not updating correctly in recipe mapping
