+++ src/domain/recipe/repositories/IRecipeRepository.ts
@@ -2,6 +2,7 @@ import type {
   MealiePaginatedRecipes,
   MealieRecipe,
   MealieCategory,
+  MealieFavoritesResponse,
   RecipeFilters,
   RecipeFormData,
   Season,
@@ -19,5 +20,8 @@ export interface IRecipeRepository {
   updateSeasons(slug: string, seasons: Season[]): Promise<MealieRecipe>
   updateCalorieTags(slug: string, calories: number): Promise<MealieRecipe>
   updateCategories(slug: string, categories: MealieCategory[]): Promise<MealieRecipe>
+  updateRating(slug: string, rating: number): Promise<void>
+  getFavorites(): Promise<MealieFavoritesResponse>
+  toggleFavorite(slug: string, isFavorite: boolean): Promise<void>
   uploadImage(slug: string, file: File): Promise<void>
 }
