+++ src/presentation/pages/RecipeDetailPage.tsx
@@ -28,6 +28,8 @@ import {
   Check,
   UtensilsCrossed,
   Sparkles,
+  Star,
+  Heart,
 } from "lucide-react"
 import { useState, useRef, useCallback, useEffect, type ChangeEvent } from "react"
 import { CookingMode } from "../components/CookingMode.tsx"
@@ -42,6 +44,10 @@ import type {
 import { SEASONS, SEASON_LABELS } from "../../shared/types/mealie.ts"
 import { getRecipeSeasonsFromTags, isSeasonTag } from "../../shared/utils/season.ts"
 import { cn } from "../../lib/utils.ts"
+
+import { useUpdateRating } from "../../presentation/hooks/useUpdateRating"
+import { useGetFavorites } from "../../presentation/hooks/useGetFavorites"
+import { useToggleFavorite } from "../../presentation/hooks/useToggleFavorite"
 
 function buildFormData(recipe: MealieRecipe): RecipeFormData {
   const structured =
@@ -59,7 +65,7 @@ function buildFormData(recipe: MealieRecipe): RecipeFormData {
         note: ing.note ?? "",
         referenceId: ing.referenceId,
       })) ?? []
-      
+
   return {
     name: recipe.name,
     description: recipe.description ?? "",
@@ -133,6 +139,8 @@ export function RecipeDetailPage() {
   const [imagePreview, setImagePreview] = useState<string | null>(null)
   const [cookingMode, setCookingMode] = useState(false)
   const fileInputRef = useRef<HTMLInputElement>(null)
+  const { getFavorites } = useGetFavorites()
+  const [ratings, setRatings] = useState<any[]>([])
 
   // Initialise formData once recipe is loaded
   useEffect(() => {
@@ -140,6 +148,12 @@ export function RecipeDetailPage() {
       setFormData(buildFormData(recipe))
       setImagePreview(recipeImageUrl(recipe, "original"))
     }
+
+    // Get favorites
+    void (async () => {
+      const data = await getFavorites()
+      setRatings(data.ratings)
+    })()
   }, [recipe, formData])
 
   const patch = useCallback((partial: Partial<RecipeFormData>) => {
@@ -163,7 +177,7 @@ export function RecipeDetailPage() {
     if (mealieUrl) setImagePreview(mealieUrl)
   }
 
-  // ─── Categories & Seasons (saved immediately, no dirty needed) ─────────────
+  // ─── Categories, Seasons, stars and favorites (saved immediately, no dirty needed) ─────────────
 
   const handleToggleCategory = async (cat: MealieCategory) => {
     if (!recipe || !formData) return
@@ -172,13 +186,18 @@ export function RecipeDetailPage() {
     const newCategories = isActive
       ? current.filter((c) => c.id !== cat.id)
       : [...current, { id: cat.id, name: cat.name, slug: cat.slug }]
-    patch({ categories: newCategories })
-    const updated = await updateCategories(recipe.slug, newCategories.map((c) => ({
-      id: c.id,
-      name: c.name,
-      slug: c.slug,
-      groupId: "",
-    })))
+    setFormData((prev) =>
+      prev ? { ...prev, categories: newCategories } : prev
+    )
+    const updated = await updateCategories(
+      recipe.slug,
+      newCategories.map((c) => ({
+        id: c.id,
+        name: c.name,
+        slug: c.slug,
+        groupId: "",
+      }))
+    )
     if (updated) setRecipe(updated)
   }
 
@@ -188,10 +207,46 @@ export function RecipeDetailPage() {
     const newSeasons = currentSeasons.includes(season)
       ? currentSeasons.filter((s) => s !== season)
       : [...currentSeasons, season]
-    patch({ seasons: newSeasons })
+    setFormData((prev) =>
+      prev ? { ...prev, seasons: newSeasons } : prev
+    )
     const updated = await updateSeasons(recipe.slug, newSeasons)
     if (updated) setRecipe(updated)
   }
+
+  const { updateRating } = useUpdateRating()
+  const handleRate = async (value: number) => {
+    if (!recipe) return
+    const success = await updateRating(recipe.slug, value)
+    if (success) {
+      setRecipe({
+        ...recipe,
+        rating: value,
+      })
+    }
+  }
+
+  const { toggleFavorite } = useToggleFavorite()
+  const [Favorite, setFavorite] = useState<boolean | null>(null)
+  // Vas vori si la recette est dans les favorie
+  const isFavorite =
+    Favorite !== null
+      ? Favorite
+      : ratings.some(
+        (r) => r.recipeId === recipe?.id && r.isFavorite
+      )
+
+  const handleToggleFavorite = async () => {
+    if (!recipe) return
+    const previous = isFavorite
+    const next = !previous
+    setFavorite(next)
+    const success = await toggleFavorite(recipe.slug, previous)
+    if (!success) {
+      setFavorite(previous)
+    }
+  }
+
 
   // ─── Ingredients ───────────────────────────────────────────────────────────
 
@@ -276,437 +331,523 @@ export function RecipeDetailPage() {
         />
       )}
 
-    <div className="mx-auto max-w-2xl space-y-6 px-4 py-6">
-      {/* Header */}
-      <div className="flex items-center justify-between">
-        <Button variant="ghost" size="sm" asChild>
-          <Link to="/recipes">&larr; Recettes</Link>
-        </Button>
+      <div className="mx-auto max-w-2xl space-y-6 px-4 py-6">
+        {/* Header */}
+        <div className="flex items-center justify-between">
+          <Button variant="ghost" size="sm" asChild>
+            <Link to="/recipes">&larr; Recettes</Link>
+          </Button>
 
-        <div className="flex items-center gap-2">
-          {recipe && (
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => setCookingMode(true)}
-              className="gap-1.5"
-            >
-              <UtensilsCrossed className="h-4 w-4" />
-              Mode cuisine
-            </Button>
-          )}
-
-          {isDirty && recipe && (
-            <>
+          <div className="flex items-center gap-2">
+            {recipe && (
               <Button
+                variant="outline"
                 size="sm"
-                variant="ghost"
-                onClick={() => {
-                  setFormData(buildFormData(recipe))
-                  setImagePreview(recipeImageUrl(recipe, "original"))
-                  setIsDirty(false)
-                }}
-                disabled={saving}
-              >
-                Annuler
-              </Button>
-              <Button
-                size="sm"
-                onClick={() => void handleSave()}
-                disabled={saving || !formData?.name.trim()}
+                onClick={() => setCookingMode(true)}
                 className="gap-1.5"
               >
-                {saving ? (
+                <UtensilsCrossed className="h-4 w-4" />
+                Mode cuisine
+              </Button>
+            )}
+
+            {isDirty && recipe && (
+              <>
+                <Button
+                  size="sm"
+                  variant="ghost"
+                  onClick={() => {
+                    setFormData(buildFormData(recipe))
+                    setImagePreview(recipeImageUrl(recipe, "original"))
+                    setIsDirty(false)
+                  }}
+                  disabled={saving}
+                >
+                  Annuler
+                </Button>
+                <Button
+                  size="sm"
+                  onClick={() => void handleSave()}
+                  disabled={saving || !formData?.name.trim()}
+                  className="gap-1.5"
+                >
+                  {saving ? (
+                    <>
+                      <Loader2 className="h-4 w-4 animate-spin" />
+                      Enregistrement…
+                    </>
+                  ) : (
+                    <>
+                      <Check className="h-4 w-4" />
+                      Enregistrer
+                    </>
+                  )}
+                </Button>
+              </>
+            )}
+          </div>
+
+        </div>
+
+        {/* Errors */}
+        {saveError && (
+          <div className="rounded-[var(--radius-xl)] border border-destructive/20 bg-destructive/8 p-4 text-sm text-destructive">
+            {saveError}
+          </div>
+        )}
+
+        {loading && <RecipeDetailSkeleton />}
+
+        {error && (
+          <div className="rounded-[var(--radius-xl)] border border-destructive/20 bg-destructive/8 p-4 text-sm text-destructive">
+            {error}
+          </div>
+        )}
+
+        {recipe && formData && (
+          <article className="space-y-6">
+            {/* Image */}
+            <div className="space-y-2">
+              <div
+                className="group relative overflow-hidden rounded-[var(--radius-xl)] cursor-pointer"
+                onClick={() => fileInputRef.current?.click()}
+                title="Cliquer pour changer la photo"
+              >
+                {imagePreview ? (
                   <>
-                    <Loader2 className="h-4 w-4 animate-spin" />
-                    Enregistrement…
+                    <img
+                      src={imagePreview}
+                      alt={recipe.name}
+                      className="aspect-video w-full object-cover"
+                    />
+                    <div className="absolute inset-0 flex items-center justify-center bg-black/0 transition-colors group-hover:bg-black/30">
+                      <span className="flex flex-col items-center gap-1.5 text-white opacity-0 transition-opacity group-hover:opacity-100">
+                        <ImagePlus className="h-7 w-7" />
+                        <span className="text-xs font-medium">Changer la photo</span>
+                      </span>
+                    </div>
                   </>
                 ) : (
-                  <>
-                    <Check className="h-4 w-4" />
-                    Enregistrer
-                  </>
-                )}
-              </Button>
-            </>
-          )}
-        </div>
-
-      </div>
-
-      {/* Errors */}
-      {saveError && (
-        <div className="rounded-[var(--radius-xl)] border border-destructive/20 bg-destructive/8 p-4 text-sm text-destructive">
-          {saveError}
-        </div>
-      )}
-
-      {loading && <RecipeDetailSkeleton />}
-
-      {error && (
-        <div className="rounded-[var(--radius-xl)] border border-destructive/20 bg-destructive/8 p-4 text-sm text-destructive">
-          {error}
-        </div>
-      )}
-
-      {recipe && formData && (
-        <article className="space-y-6">
-          {/* Image */}
-          <div className="space-y-2">
-            <div
-              className="group relative overflow-hidden rounded-[var(--radius-xl)] cursor-pointer"
-              onClick={() => fileInputRef.current?.click()}
-              title="Cliquer pour changer la photo"
-            >
-              {imagePreview ? (
-                <>
-                  <img
-                    src={imagePreview}
-                    alt={recipe.name}
-                    className="aspect-video w-full object-cover"
-                  />
-                  <div className="absolute inset-0 flex items-center justify-center bg-black/0 transition-colors group-hover:bg-black/30">
-                    <span className="flex flex-col items-center gap-1.5 text-white opacity-0 transition-opacity group-hover:opacity-100">
-                      <ImagePlus className="h-7 w-7" />
-                      <span className="text-xs font-medium">Changer la photo</span>
-                    </span>
+                  <div className="flex aspect-video w-full flex-col items-center justify-center gap-2 rounded-[var(--radius-xl)] border-2 border-dashed border-border bg-muted/30 text-muted-foreground transition-colors hover:border-ring hover:bg-muted/50">
+                    <ImagePlus className="h-8 w-8" />
+                    <span className="text-sm">Cliquer pour ajouter une photo</span>
                   </div>
-                </>
-              ) : (
-                <div className="flex aspect-video w-full flex-col items-center justify-center gap-2 rounded-[var(--radius-xl)] border-2 border-dashed border-border bg-muted/30 text-muted-foreground transition-colors hover:border-ring hover:bg-muted/50">
-                  <ImagePlus className="h-8 w-8" />
-                  <span className="text-sm">Cliquer pour ajouter une photo</span>
+                )}
+              </div>
+              <input
+                ref={fileInputRef}
+                type="file"
+                accept="image/*"
+                className="hidden"
+                onChange={handleImageChange}
+              />
+
+              {/* Bouton photo via IA — WIP */}
+              <div className="flex items-center gap-2 flex-wrap">
+                <span title="WIP — fonctionnalité en cours de développement">
+                  <Button
+                    type="button"
+                    variant="outline"
+                    size="sm"
+                    disabled
+                    onClick={() => void handleAiImage()}
+                    className="gap-1.5 cursor-not-allowed"
+                  >
+                    <Sparkles className="h-4 w-4" />
+                    Photo via IA
+                  </Button>
+                </span>
+                <select
+                  value={aiProvider}
+                  onChange={(e) => setAiProvider(e.target.value as ImageProvider)}
+                  disabled
+                  className="h-8 rounded-md border border-input bg-background px-2 text-xs text-foreground opacity-50 cursor-not-allowed"
+                >
+                  {IMAGE_PROVIDERS.map((p) => (
+                    <option key={p.value} value={p.value}>{p.label}</option>
+                  ))}
+                </select>
+              </div>
+            </div>
+
+            {/* Titre */}
+            <div className="space-y-3">
+              <InlineEditText
+                value={formData.name}
+                displayValue={
+                  <span className="font-heading text-2xl font-bold leading-snug tracking-tight">
+                    {formData.name}
+                  </span>
+                }
+                onChange={(v) => patch({ name: v })}
+                placeholder="Nom de la recette"
+                as="h1"
+                inputClassName="font-heading text-2xl font-bold"
+                disabled={saving}
+              />
+
+              {/* Catégories */}
+              {allCategories.length > 0 && (
+                <div className="flex flex-wrap gap-1.5">
+                  {allCategories.map((cat) => {
+                    const active = formData.categories.some((c) => c.id === cat.id)
+                    return (
+                      <Badge
+                        key={cat.id}
+                        variant={active ? "default" : "outline"}
+                        className="cursor-pointer select-none transition-colors text-xs"
+                        onClick={() => void handleToggleCategory(cat)}
+                      >
+                        {cat.name}
+                      </Badge>
+                    )
+                  })}
                 </div>
               )}
-            </div>
-            <input
-              ref={fileInputRef}
-              type="file"
-              accept="image/*"
-              className="hidden"
-              onChange={handleImageChange}
-            />
 
-            {/* Bouton photo via IA — WIP */}
-            <div className="flex items-center gap-2 flex-wrap">
-              <span title="WIP — fonctionnalité en cours de développement">
+              {/* Saisons */}
+              <div className="space-y-1.5">
+                {formData.seasons.length > 0 && (
+                  <div className="flex flex-wrap gap-1.5">
+                    {formData.seasons.map((season) => (
+                      <SeasonBadge key={season} season={season} size="md" />
+                    ))}
+                  </div>
+                )}
+                <div className="flex flex-wrap gap-1.5">
+                  {SEASONS.filter((s) => s !== "sans").map((season: Season) => {
+                    const active = formData.seasons.includes(season)
+                    return (
+                      <Badge
+                        key={season}
+                        variant={active ? "default" : "outline"}
+                        className="cursor-pointer select-none transition-colors text-xs"
+                        onClick={() => void handleToggleSeason(season)}
+                      >
+                        {SEASON_LABELS[season]}
+                      </Badge>
+                    )
+                  })}
+                </div>
+              </div>
+
+              {/* Durées */}
+              <div className="flex flex-wrap gap-4">
+                <InlineEditDuration
+                  label="Préparation"
+                  value={formData.prepTime}
+                  displayRaw={recipe.prepTime}
+                  onChange={(v) => patch({ prepTime: v })}
+                  disabled={saving}
+                />
+                <InlineEditDuration
+                  label="Cuisson"
+                  value={formData.performTime}
+                  displayRaw={recipe.performTime}
+                  onChange={(v) => patch({ performTime: v })}
+                  disabled={saving}
+                />
+                <InlineEditDuration
+                  label="Total"
+                  value={formData.totalTime}
+                  displayRaw={recipe.totalTime}
+                  onChange={(v) => patch({ totalTime: v })}
+                  disabled={saving}
+                />
+              </div>
+
+              {/* Rating + Favorite */}
+              <div className="flex items-center justify-between">
+
+                {/* Rating */}
+                <div className="flex items-center gap-1.5">
+                  {Array.from({ length: 5 }).map((_, i) => {
+                    const value = i + 1
+                    const current = recipe?.rating ?? 0
+
+                    return (
+                      <Star
+                        key={i}
+                        onClick={() => void handleRate(value)}
+                        className={cn(
+                          "h-3.5 w-3.5 cursor-pointer transition",
+                          loading && "pointer-events-none opacity-50",
+                          value <= current
+                            ? "text-yellow-400 fill-yellow-400"
+                            : "text-muted-foreground/30 hover:text-yellow-300"
+                        )}
+                      />
+                    )
+                  })}
+                </div>
+
+                {/* Favorite ❤️ */}
+                <button
+                  onClick={() => void handleToggleFavorite()}
+                  className={cn(
+                    "ml-2 transition-all duration-150 hover:scale-110",
+                    isFavorite ? "text-red-500" : "text-muted-foreground/40 hover:text-red-400"
+                  )}
+                >
+                  <Heart
+                    className={cn(
+                      "h-5 w-5 transition-all",
+                      isFavorite && "fill-red-500 text-red-500"
+                    )}
+                  />
+                </button>
+              </div>
+
+              {/* nutrition / Pas encore editable */}
+              {recipe?.nutrition?.calories && (
+                <div
+                  className={cn(
+                    "space-y-2.5 rounded-[var(--radius-xl)]",
+                    "border border-border/50 bg-secondary/30 p-3.5",
+                  )}
+                >
+                  <p className="text-[10px] font-bold uppercase tracking-[0.10em] text-muted-foreground/60">
+                    Nutrition
+                  </p>
+
+                  <div className="flex flex-wrap gap-1.5">
+                    <Badge variant="outline">
+                      {recipe.nutrition.calories} kcal
+                    </Badge>
+
+                    {recipe.nutrition.proteinContent && (
+                      <Badge variant="outline">
+                        {recipe.nutrition.proteinContent}g protéines
+                      </Badge>
+                    )}
+
+                    {recipe.nutrition.carbohydrateContent && (
+                      <Badge variant="outline">
+                        {recipe.nutrition.carbohydrateContent}g glucides
+                      </Badge>
+                    )}
+
+                    {recipe.nutrition.fatContent && (
+                      <Badge variant="outline">
+                        {recipe.nutrition.fatContent}g lipides
+                      </Badge>
+                    )}
+
+                    {recipe.nutrition.fiberContent && (
+                      <Badge variant="outline">
+                        {recipe.nutrition.fiberContent}g fibres
+                      </Badge>
+                    )}
+                  </div>
+                </div>
+              )}
+
+              {/* Description */}
+              <InlineEditText
+                value={formData.description}
+                onChange={(v) => patch({ description: v })}
+                placeholder="Ajouter une description…"
+                multiline
+                rows={3}
+                as="p"
+                className="text-sm text-muted-foreground leading-relaxed"
+                disabled={saving}
+              />
+            </div>
+
+            {/* Ingrédients */}
+            <section className="space-y-3">
+              <div className="flex items-center justify-between">
+                <h2 className="font-heading text-lg font-bold tracking-tight">Ingrédients</h2>
                 <Button
                   type="button"
                   variant="outline"
                   size="sm"
-                  disabled
-                  onClick={() => void handleAiImage()}
-                  className="gap-1.5 cursor-not-allowed"
+                  onClick={addIngredient}
+                  disabled={saving}
+                  className="gap-1.5"
                 >
-                  <Sparkles className="h-4 w-4" />
-                  Photo via IA
+                  <Plus className="h-3.5 w-3.5" />
+                  Ajouter
                 </Button>
-              </span>
-              <select
-                value={aiProvider}
-                onChange={(e) => setAiProvider(e.target.value as ImageProvider)}
-                disabled
-                className="h-8 rounded-md border border-input bg-background px-2 text-xs text-foreground opacity-50 cursor-not-allowed"
-              >
-                {IMAGE_PROVIDERS.map((p) => (
-                  <option key={p.value} value={p.value}>{p.label}</option>
-                ))}
-              </select>
-            </div>
-          </div>
+              </div>
 
-          {/* Titre */}
-          <div className="space-y-3">
-            <InlineEditText
-              value={formData.name}
-              displayValue={
-                <span className="font-heading text-2xl font-bold leading-snug tracking-tight">
-                  {formData.name}
-                </span>
-              }
-              onChange={(v) => patch({ name: v })}
-              placeholder="Nom de la recette"
-              as="h1"
-              inputClassName="font-heading text-2xl font-bold"
-              disabled={saving}
-            />
+              <div className="hidden sm:grid sm:grid-cols-[16px_70px_1fr_1fr_1.5fr_32px] sm:gap-2 sm:items-center px-1">
+                <span />
+                <span className="text-xs text-muted-foreground font-medium">Qté</span>
+                <span className="text-xs text-muted-foreground font-medium">Unité</span>
+                <span className="text-xs text-muted-foreground font-medium">Ingrédient</span>
+                <span className="text-xs text-muted-foreground font-medium">Notes</span>
+                <span />
+              </div>
 
-            {/* Catégories */}
-            {allCategories.length > 0 && (
-              <div className="flex flex-wrap gap-1.5">
-                {allCategories.map((cat) => {
-                  const active = formData.categories.some((c) => c.id === cat.id)
-                  return (
-                    <Badge
-                      key={cat.id}
-                      variant={active ? "default" : "outline"}
-                      className="cursor-pointer select-none transition-colors text-xs"
-                      onClick={() => void handleToggleCategory(cat)}
+              <div className="space-y-2">
+                {formData.recipeIngredient.map((ing, index) => (
+                  <div
+                    key={index}
+                    className="grid grid-cols-[16px_55px_1fr_1fr_32px] sm:grid-cols-[16px_70px_1fr_1fr_1.5fr_32px] gap-2 items-center"
+                  >
+                    <GripVertical className="h-4 w-4 text-muted-foreground" />
+
+                    <Input
+                      type="text"
+                      inputMode="decimal"
+                      placeholder="Qté"
+                      value={ing.quantity}
+                      onChange={(e) => updateIngredientField(index, { quantity: e.target.value })}
+                      disabled={saving}
+                      className="min-w-0 px-2"
+                      aria-label={`Quantité ingrédient ${index + 1}`}
+                    />
+
+                    <Autocomplete
+                      value={ing.unit}
+                      onChange={(value, option) =>
+                        updateIngredientField(index, { unit: value, unitId: option?.id })
+                      }
+                      options={unitOptions}
+                      placeholder="Unité…"
+                      disabled={saving}
+                      inputClassName="bg-white dark:bg-zinc-900"
+                      aria-label={`Unité ingrédient ${index + 1}`}
+                    />
+
+                    <Autocomplete
+                      value={ing.food}
+                      onChange={(value, option) =>
+                        updateIngredientField(index, {
+                          food: value,
+                          foodId: option && option.id !== "__create__" ? option.id : undefined,
+                        })
+                      }
+                      options={foodOptions}
+                      placeholder="Ingrédient…"
+                      disabled={saving}
+                      allowCreate
+                      createLabel={(v) => `Créer "${v}"`}
+                      inputClassName="bg-white dark:bg-zinc-900"
+                      aria-label={`Ingrédient ${index + 1}`}
+                    />
+
+                    <Input
+                      type="text"
+                      placeholder="Notes…"
+                      value={ing.note}
+                      onChange={(e) => updateIngredientField(index, { note: e.target.value })}
+                      disabled={saving}
+                      className="hidden sm:block min-w-0 px-2"
+                      aria-label={`Notes ingrédient ${index + 1}`}
+                    />
+
+                    <Button
+                      type="button"
+                      variant="ghost"
+                      size="icon"
+                      onClick={() => removeIngredient(index)}
+                      disabled={saving || formData.recipeIngredient.length <= 1}
+                      className="h-8 w-8 text-muted-foreground hover:text-destructive"
+                      aria-label={`Supprimer ingrédient ${index + 1}`}
                     >
-                      {cat.name}
-                    </Badge>
-                  )
-                })}
+                      <Trash2 className="h-3.5 w-3.5" />
+                    </Button>
+                  </div>
+                ))}
+              </div>
+            </section>
+
+            {/* Instructions */}
+            <section className="space-y-3">
+              <div className="flex items-center justify-between">
+                <h2 className="font-heading text-lg font-bold tracking-tight">Instructions</h2>
+                <Button
+                  type="button"
+                  variant="outline"
+                  size="sm"
+                  onClick={addInstruction}
+                  disabled={saving}
+                  className="gap-1.5"
+                >
+                  <Plus className="h-3.5 w-3.5" />
+                  Ajouter
+                </Button>
+              </div>
+
+              <ol className="space-y-4">
+                {formData.recipeInstructions.map((step, index) => (
+                  <li key={index} className="flex gap-3 items-start">
+                    <span className="flex h-6 w-6 shrink-0 items-center justify-center rounded-full bg-primary/8 text-[11px] font-bold text-primary mt-1">
+                      {index + 1}
+                    </span>
+                    <div className="flex-1 min-w-0">
+                      <textarea
+                        value={step.text}
+                        onChange={(e) => updateInstruction(index, e.target.value)}
+                        placeholder={`Étape ${index + 1}…`}
+                        disabled={saving}
+                        rows={2}
+                        className={cn(
+                          "w-full rounded-md border-transparent bg-transparent px-2 py-1 text-sm text-muted-foreground leading-relaxed",
+                          "placeholder:text-muted-foreground/50 focus-visible:outline-none focus-visible:border-input",
+                          "focus-visible:bg-background focus-visible:ring-1 focus-visible:ring-ring",
+                          "hover:bg-muted/30 transition-colors resize-none",
+                          "disabled:cursor-not-allowed disabled:opacity-50",
+                        )}
+                        aria-label={`Étape ${index + 1}`}
+                      />
+                    </div>
+                    <Button
+                      type="button"
+                      variant="ghost"
+                      size="icon"
+                      onClick={() => removeInstruction(index)}
+                      disabled={saving}
+                      className="h-8 w-8 text-muted-foreground hover:text-destructive shrink-0 mt-0.5"
+                    >
+                      <Trash2 className="h-3.5 w-3.5" />
+                    </Button>
+                  </li>
+                ))}
+              </ol>
+            </section>
+
+            {/* Sticky save bar when dirty */}
+            {isDirty && (
+              <div className="sticky bottom-4 flex justify-end gap-2 rounded-[var(--radius-xl)] border border-border bg-background/95 px-4 py-3 shadow-lg backdrop-blur-sm">
+                <Button
+                  variant="outline"
+                  size="sm"
+                  onClick={() => {
+                    setFormData(buildFormData(recipe))
+                    setImagePreview(recipeImageUrl(recipe, "original"))
+                    setIsDirty(false)
+                  }}
+                  disabled={saving}
+                >
+                  Annuler
+                </Button>
+                <Button
+                  size="sm"
+                  onClick={() => void handleSave()}
+                  disabled={saving || !formData.name.trim()}
+                  className="gap-1.5"
+                >
+                  {saving ? (
+                    <>
+                      <Loader2 className="h-4 w-4 animate-spin" />
+                      Enregistrement…
+                    </>
+                  ) : (
+                    <>
+                      <Check className="h-4 w-4" />
+                      Enregistrer les modifications
+                    </>
+                  )}
+                </Button>
               </div>
             )}
-
-            {/* Saisons */}
-            <div className="space-y-1.5">
-              {formData.seasons.length > 0 && (
-                <div className="flex flex-wrap gap-1.5">
-                  {formData.seasons.map((season) => (
-                    <SeasonBadge key={season} season={season} size="md" />
-                  ))}
-                </div>
-              )}
-              <div className="flex flex-wrap gap-1.5">
-                {SEASONS.filter((s) => s !== "sans").map((season: Season) => {
-                  const active = formData.seasons.includes(season)
-                  return (
-                    <Badge
-                      key={season}
-                      variant={active ? "default" : "outline"}
-                      className="cursor-pointer select-none transition-colors text-xs"
-                      onClick={() => void handleToggleSeason(season)}
-                    >
-                      {SEASON_LABELS[season]}
-                    </Badge>
-                  )
-                })}
-              </div>
-            </div>
-
-            {/* Durées */}
-            <div className="flex flex-wrap gap-4">
-              <InlineEditDuration
-                label="Préparation"
-                value={formData.prepTime}
-                displayRaw={recipe.prepTime}
-                onChange={(v) => patch({ prepTime: v })}
-                disabled={saving}
-              />
-              <InlineEditDuration
-                label="Cuisson"
-                value={formData.performTime}
-                displayRaw={recipe.performTime}
-                onChange={(v) => patch({ performTime: v })}
-                disabled={saving}
-              />
-              <InlineEditDuration
-                label="Total"
-                value={formData.totalTime}
-                displayRaw={recipe.totalTime}
-                onChange={(v) => patch({ totalTime: v })}
-                disabled={saving}
-              />
-            </div>
-
-            {/* Description */}
-            <InlineEditText
-              value={formData.description}
-              onChange={(v) => patch({ description: v })}
-              placeholder="Ajouter une description…"
-              multiline
-              rows={3}
-              as="p"
-              className="text-sm text-muted-foreground leading-relaxed"
-              disabled={saving}
-            />
-          </div>
-
-          {/* Ingrédients */}
-          <section className="space-y-3">
-            <div className="flex items-center justify-between">
-              <h2 className="font-heading text-lg font-bold tracking-tight">Ingrédients</h2>
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                onClick={addIngredient}
-                disabled={saving}
-                className="gap-1.5"
-              >
-                <Plus className="h-3.5 w-3.5" />
-                Ajouter
-              </Button>
-            </div>
-
-            <div className="hidden sm:grid sm:grid-cols-[16px_70px_1fr_1fr_1.5fr_32px] sm:gap-2 sm:items-center px-1">
-              <span />
-              <span className="text-xs text-muted-foreground font-medium">Qté</span>
-              <span className="text-xs text-muted-foreground font-medium">Unité</span>
-              <span className="text-xs text-muted-foreground font-medium">Ingrédient</span>
-              <span className="text-xs text-muted-foreground font-medium">Notes</span>
-              <span />
-            </div>
-
-            <div className="space-y-2">
-              {formData.recipeIngredient.map((ing, index) => (
-                <div
-                  key={index}
-                  className="grid grid-cols-[16px_55px_1fr_1fr_32px] sm:grid-cols-[16px_70px_1fr_1fr_1.5fr_32px] gap-2 items-center"
-                >
-                  <GripVertical className="h-4 w-4 text-muted-foreground" />
-
-                  <Input
-                    type="text"
-                    inputMode="decimal"
-                    placeholder="Qté"
-                    value={ing.quantity}
-                    onChange={(e) => updateIngredientField(index, { quantity: e.target.value })}
-                    disabled={saving}
-                    className="min-w-0 px-2"
-                    aria-label={`Quantité ingrédient ${index + 1}`}
-                  />
-
-                  <Autocomplete
-                    value={ing.unit}
-                    onChange={(value, option) =>
-                      updateIngredientField(index, { unit: value, unitId: option?.id })
-                    }
-                    options={unitOptions}
-                    placeholder="Unité…"
-                    disabled={saving}
-                    inputClassName="bg-white dark:bg-zinc-900"
-                    aria-label={`Unité ingrédient ${index + 1}`}
-                  />
-
-                  <Autocomplete
-                    value={ing.food}
-                    onChange={(value, option) =>
-                      updateIngredientField(index, {
-                        food: value,
-                        foodId: option && option.id !== "__create__" ? option.id : undefined,
-                      })
-                    }
-                    options={foodOptions}
-                    placeholder="Ingrédient…"
-                    disabled={saving}
-                    allowCreate
-                    createLabel={(v) => `Créer "${v}"`}
-                    inputClassName="bg-white dark:bg-zinc-900"
-                    aria-label={`Ingrédient ${index + 1}`}
-                  />
-
-                  <Input
-                    type="text"
-                    placeholder="Notes…"
-                    value={ing.note}
-                    onChange={(e) => updateIngredientField(index, { note: e.target.value })}
-                    disabled={saving}
-                    className="hidden sm:block min-w-0 px-2"
-                    aria-label={`Notes ingrédient ${index + 1}`}
-                  />
-
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="icon"
-                    onClick={() => removeIngredient(index)}
-                    disabled={saving || formData.recipeIngredient.length <= 1}
-                    className="h-8 w-8 text-muted-foreground hover:text-destructive"
-                    aria-label={`Supprimer ingrédient ${index + 1}`}
-                  >
-                    <Trash2 className="h-3.5 w-3.5" />
-                  </Button>
-                </div>
-              ))}
-            </div>
-          </section>
-
-          {/* Instructions */}
-          <section className="space-y-3">
-            <div className="flex items-center justify-between">
-              <h2 className="font-heading text-lg font-bold tracking-tight">Instructions</h2>
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                onClick={addInstruction}
-                disabled={saving}
-                className="gap-1.5"
-              >
-                <Plus className="h-3.5 w-3.5" />
-                Ajouter
-              </Button>
-            </div>
-
-            <ol className="space-y-4">
-              {formData.recipeInstructions.map((step, index) => (
-                <li key={index} className="flex gap-3 items-start">
-                  <span className="flex h-6 w-6 shrink-0 items-center justify-center rounded-full bg-primary/8 text-[11px] font-bold text-primary mt-1">
-                    {index + 1}
-                  </span>
-                  <div className="flex-1 min-w-0">
-                    <textarea
-                      value={step.text}
-                      onChange={(e) => updateInstruction(index, e.target.value)}
-                      placeholder={`Étape ${index + 1}…`}
-                      disabled={saving}
-                      rows={2}
-                      className={cn(
-                        "w-full rounded-md border-transparent bg-transparent px-2 py-1 text-sm text-muted-foreground leading-relaxed",
-                        "placeholder:text-muted-foreground/50 focus-visible:outline-none focus-visible:border-input",
-                        "focus-visible:bg-background focus-visible:ring-1 focus-visible:ring-ring",
-                        "hover:bg-muted/30 transition-colors resize-none",
-                        "disabled:cursor-not-allowed disabled:opacity-50",
-                      )}
-                      aria-label={`Étape ${index + 1}`}
-                    />
-                  </div>
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="icon"
-                    onClick={() => removeInstruction(index)}
-                    disabled={saving}
-                    className="h-8 w-8 text-muted-foreground hover:text-destructive shrink-0 mt-0.5"
-                  >
-                    <Trash2 className="h-3.5 w-3.5" />
-                  </Button>
-                </li>
-              ))}
-            </ol>
-          </section>
-
-          {/* Sticky save bar when dirty */}
-          {isDirty && (
-            <div className="sticky bottom-4 flex justify-end gap-2 rounded-[var(--radius-xl)] border border-border bg-background/95 px-4 py-3 shadow-lg backdrop-blur-sm">
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => {
-                  setFormData(buildFormData(recipe))
-                  setImagePreview(recipeImageUrl(recipe, "original"))
-                  setIsDirty(false)
-                }}
-                disabled={saving}
-              >
-                Annuler
-              </Button>
-              <Button
-                size="sm"
-                onClick={() => void handleSave()}
-                disabled={saving || !formData.name.trim()}
-                className="gap-1.5"
-              >
-                {saving ? (
-                  <>
-                    <Loader2 className="h-4 w-4 animate-spin" />
-                    Enregistrement…
-                  </>
-                ) : (
-                  <>
-                    <Check className="h-4 w-4" />
-                    Enregistrer les modifications
-                  </>
-                )}
-              </Button>
-            </div>
-          )}
-        </article>
-      )}
-    </div>
+          </article>
+        )}
+      </div>
     </>
   )
 }
